fix: refactor state_migration to expose single pub api
