+++ blockchain/state_manager/src/lib.rs
@@ -5,14 +5,14 @@ pub mod chain_rand;
 mod errors;
 mod metrics;
 mod utils;
-pub use utils::{is_valid_for_sending, reveal_five_trees};
+use forest_state_migration::run_state_migrations;
+pub use utils::is_valid_for_sending;
 
 mod vm_circ_supply;
 
 use std::{num::NonZeroUsize, sync::Arc};
 
 use ahash::{HashMap, HashMapExt};
-use anyhow::bail;
 use chain_rand::ChainRand;
 use cid::Cid;
 use fil_actor_interface::*;
@@ -23,7 +23,7 @@ use forest_chain::{ChainStore, HeadChange};
 use forest_interpreter::{resolve_to_key_addr, BlockMessages, RewardCalc, VM};
 use forest_json::message_receipt;
 use forest_message::{ChainMessage, Message as MessageTrait};
-use forest_networks::{ChainConfig, Height};
+use forest_networks::ChainConfig;
 use forest_shim::{
     address::{Address, Payload, Protocol, BLS_PUB_LEN},
     econ::TokenAmount,
@@ -46,10 +46,7 @@ use num_traits::identities::Zero;
 use once_cell::unsync::Lazy;
 use parking_lot::Mutex as SyncMutex;
 use serde::{Deserialize, Serialize};
-use tokio::{
-    sync::{broadcast::error::RecvError, Mutex as TokioMutex, RwLock},
-    time,
-};
+use tokio::sync::{broadcast::error::RecvError, Mutex as TokioMutex, RwLock};
 use tracing::{debug, error, info, instrument, trace, warn};
 use vm_circ_supply::GenesisInfo;
 
@@ -447,45 +444,7 @@ where
         parent_state: &Cid,
         epoch: ChainEpoch,
     ) -> anyhow::Result<Option<Cid>> {
-        match epoch {
-            x if x == self.chain_config.epoch(Height::Hygge) => {
-                info!("Running Hygge migration at epoch {epoch}");
-                let start_time = time::Instant::now();
-                let new_state = forest_state_migration::run_nv18_migration(
-                    &self.chain_config,
-                    self.blockstore(),
-                    parent_state,
-                    epoch,
-                )?;
-                let elapsed = start_time.elapsed().as_secs_f32();
-                if new_state != *parent_state {
-                    reveal_five_trees();
-                    info!("State migration successful, took: {elapsed}s");
-                } else {
-                    bail!("State post migration must not match. Previous state: {parent_state}, new state: {new_state}. Took {elapsed}s");
-                }
-                Ok(Some(new_state))
-            }
-            x if x == self.chain_config.epoch(Height::Lightning) => {
-                info!("Running Lightning migration at epoch {epoch}");
-                let start_time = time::Instant::now();
-                let new_state = forest_state_migration::run_nv19_migration(
-                    &self.chain_config,
-                    self.blockstore(),
-                    parent_state,
-                    epoch,
-                )?;
-                let elapsed = start_time.elapsed().as_secs_f32();
-                if new_state != *parent_state {
-                    reveal_five_trees();
-                    info!("State migration successful, took: {elapsed}s");
-                } else {
-                    bail!("State post migration must not match. Previous state: {parent_state}, new state: {new_state}. Took {elapsed}s");
-                }
-                Ok(Some(new_state))
-            }
-            _ => Ok(None),
-        }
+        run_state_migrations(epoch, self.chain_config(), self.blockstore(), parent_state)
     }
 
     /// Returns the pair of (parent state root, message receipt root). This will
