+++ blockchain/state_manager/src/utils.rs
@@ -15,7 +15,6 @@ use forest_utils::encoding::prover_id_from_u64;
 use fvm_ipld_bitfield::BitField;
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::bytes_32;
-use tracing::info;
 
 use crate::{errors::*, StateManager};
 
@@ -155,37 +154,6 @@ pub fn is_valid_for_sending(network_version: NetworkVersion, actor: &ActorState)
     } else {
         false
     };
-}
-
-/// Reveals five trees arranged in an order that resemble the forest logo.
-/// To be used at anyone's convenience.
-pub fn reveal_five_trees() {
-    info!(
-        r###"
-
-                                        ▄
-                                       ██
-                                      ████
-                         █           ▄█████           ▄▄
-                        ███         ▐██████▌         ▄██▄
-                       █████        ████████▄       ▄████▄
-          █▄         ▄███████      ██████████      ▄██████▄          █▄
-        ▄███▄        █████████    ████████████    ▄████████▄       ▄████
-       ███████     ▄███████████  ▐█████████████  ███████████▄     ███████▄
-     ▄█████████▄  ▄████████████ ▐██████████████▌ ████████████▄  ▄█████████▄
-    ████████████ ▄████████████▀ ████████████████▄ ████████████▄ ████████████
-  ▄████████████ ▄████████████▌ ██████████████████ ▐████████████▄ ████████████▄
- ▄████████████ ▄█████████████ ████████████████████ ▀████████████▄ █████████████
-▀█████████████ ▀████████████▄ ▀██████████████████▀ ▄████████████▀ ▄████████████▀
-   ▀████████████▄ ▀████████████  ▀████████████▀▀ ████████████▀  ████████████▀
-      ▀███████▀      ▀███████▀      ▀██████▀▀     ▀▀██████▀      ▀▀██████▀
-         ▀█▀            ██▀            ██▀           ▀██            ▀██
-         ▐█▌            ▐█             ▐█             █▌             █▌
-         ▐█▌            ▐█             ▐█             █▌             █▌
-         ▐█▌            ▐█             ▐█             █▌             █▌
-
-"###
-    );
 }
 
 /// Generates sector challenge indexes for use in winning PoSt verification.
