+++ utils/forest_utils/Cargo.toml
@@ -44,6 +44,7 @@ serde_yaml.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["fs"] }
 toml.workspace = true
+tracing.workspace = true
 url.workspace = true
 
 [dev-dependencies]
