+++ utils/forest_utils/src/miscs/logo.rs
@@ -1,0 +1,35 @@
+// Copyright 2019-2023 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use tracing::info;
+
+/// Reveals five trees arranged in an order that resemble the forest logo.
+/// To be used at anyone's convenience.
+pub fn reveal_five_trees() {
+    info!(
+        r###"
+
+                                        ▄
+                                       ██
+                                      ████
+                         █           ▄█████           ▄▄
+                        ███         ▐██████▌         ▄██▄
+                       █████        ████████▄       ▄████▄
+          █▄         ▄███████      ██████████      ▄██████▄          █▄
+        ▄███▄        █████████    ████████████    ▄████████▄       ▄████
+       ███████     ▄███████████  ▐█████████████  ███████████▄     ███████▄
+     ▄█████████▄  ▄████████████ ▐██████████████▌ ████████████▄  ▄█████████▄
+    ████████████ ▄████████████▀ ████████████████▄ ████████████▄ ████████████
+  ▄████████████ ▄████████████▌ ██████████████████ ▐████████████▄ ████████████▄
+ ▄████████████ ▄█████████████ ████████████████████ ▀████████████▄ █████████████
+▀█████████████ ▀████████████▄ ▀██████████████████▀ ▄████████████▀ ▄████████████▀
+   ▀████████████▄ ▀████████████  ▀████████████▀▀ ████████████▀  ████████████▀
+      ▀███████▀      ▀███████▀      ▀██████▀▀     ▀▀██████▀      ▀▀██████▀
+         ▀█▀            ██▀            ██▀           ▀██            ▀██
+         ▐█▌            ▐█             ▐█             █▌             █▌
+         ▐█▌            ▐█             ▐█             █▌             █▌
+         ▐█▌            ▐█             ▐█             █▌             █▌
+
+"###
+    );
+}
