+++ utils/forest_utils/src/miscs/mod.rs
@@ -3,3 +3,5 @@
 
 mod either;
 pub use either::*;
+mod logo;
+pub use logo::*;
